refactor(core/stellar): move generic Soroban writers to writers.py.
SCVal, SCAddress and authorized-invocation writers are used by both the
transaction flow and the standalone authorization signing flow, so they
are no longer operation-specific.

[no changelog]
